정산 데이터 처리2

+++ 3Week_Mission/src/main/java/com/ll/exam/final__2022_10_08/app/rebate/entity/RebateOrderItem.java
@@ -44,12 +44,12 @@ public class RebateOrderItem extends BaseEntity {
     private CashLog rebateCashLog;
 
     // 가격
-    private long price; // 권장판매가
-    private long salePrice; // 실제판매가
-    private long wholesalePrice; // 도매가
-    private long pgFee; // 결제대행사 수수료
-    private long payPrice; // 결제금액
-    private long refundPrice; // 환불금액
+    private int price; // 권장판매가
+    private int salePrice; // 실제판매가
+    private int wholesalePrice; // 도매가
+    private int pgFee; // 결제대행사 수수료
+    private int payPrice; // 결제금액
+    private int refundPrice; // 환불금액
     private boolean isPaid; // 결제여부
     private LocalDateTime payDate; // 결제날짜
 
@@ -58,7 +58,7 @@ public class RebateOrderItem extends BaseEntity {
 
     private String sellerName;
 
-    private double calculateRebatePrice;
+    private int calculateRebatePrice;
     // 주문품목
     private LocalDateTime orderItemCreateDate;
     private boolean rebateAvailable;
@@ -76,7 +76,7 @@ public class RebateOrderItem extends BaseEntity {
         isPaid = orderItem.isPaid();
         payDate = orderItem.getPayDate();
         sellerName = orderItem.getProduct().getAuthor().getName();
-        calculateRebatePrice = orderItem.getPayPrice() * 0.5;
+        calculateRebatePrice = (int) (orderItem.getPayPrice() * 0.5);
         // 상품 추가데이터
         productSubject = orderItem.getProduct().getSubject();
